requirements.txt
pyyaml
six
river-admin==0.6.0
django-river
django==2.2.13
